+++ CHANGELOG.md
@@ -3,5 +3,6 @@
 ## 0.3.0/main - YYYY-MM-DD
 
 - Refactor default generation command to own `generate` command. Added command to list supported containers.
+- New reqquired argument for generation command: `--containers`
 - Container has now it's own struct which allows to support more containers
 - Added Dependabot config